# Salus/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salus API",
        "description": "Documentation de l'API Salus.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Lister les rendez-vous",
                "operationId": "50f9b9ffc54ca980703668ea12f9e899",
                "responses": {
                    "200": {
                        "description": "Liste des rendez-vous",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des rendez-vous"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Appointment"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Creer un rendez-vous",
                "operationId": "165597a3b725c7acba4aa39883e6705e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "doctor_id",
                                    "appointment_date"
                                ],
                                "properties": {
                                    "doctor_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "appointment_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-04-02T09:30:00Z"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Premiere consultation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rendez-vous cree",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rendez-vous cree"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Appointment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    },
                    "422": {
                        "description": "Validation echouee"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/appointments/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Voir un rendez-vous",
                "operationId": "429b060de73d97219ad830c5aa0cb113",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail du rendez-vous",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detail du rendez-vous"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Appointment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    },
                    "404": {
                        "description": "Rendez-vous introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Mettre a jour un rendez-vous",
                "operationId": "33a82366e2792c7b662658754abf476c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "doctor_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "appointment_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-04-02T09:30:00Z"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "confirmed"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Confirme par le medecin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rendez-vous mis a jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rendez-vous mis a jour"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Appointment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    },
                    "404": {
                        "description": "Rendez-vous introuvable"
                    },
                    "422": {
                        "description": "Validation echouee"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Annuler un rendez-vous",
                "operationId": "765cea908af4fdcde0b04395d3c75188",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rendez-vous annule",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rendez-vous annule"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Appointment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    },
                    "404": {
                        "description": "Rendez-vous introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Doctor": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Dr. Amina El Idrissi"
                    },
                    "specialty": {
                        "type": "string",
                        "example": "Cardiologie"
                    },
                    "city": {
                        "type": "string",
                        "example": "Casablanca"
                    },
                    "consultation_price": {
                        "type": "string",
                        "example": "350.00"
                    },
                    "available_days": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Monday"
                        }
                    },
                    "yearsofexperience": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
            "Appointment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "doctor_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "appointment_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-02T09:30:00Z"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Premiere consultation"
                    },
                    "doctor": {
                        "$ref": "#/components/schemas/Doctor"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Entrez le token au format: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}